+++ src/config.php
@@ -167,7 +167,7 @@ return [
         'dropdown.items' => Components\Dropdown\Items::class,
         'error' => Components\Form\Error::class,
         'errors' => Components\Errors::class,
-        'fileupload' => Components\Form\FileUpload::class,
+        'upload' => Components\Form\Upload::class,
         'hint' => Components\Form\Hint::class,
         'icon' => Components\Icon::class,
         'input' => Components\Form\Input::class,
